It is really hard to get into X and XCB
Need to figure out how to use xcb before I'm able to generate rust bindings
for it.

Why generate my own when there is an xcb rust library? Because if I can generate
my own then I'm able to assert whether or not that library is properly built.
With the experience of diving deep into these things I'll be better able to help
developing bindings and abstractions for rust.